refactor(memory): extract document render and budget pruning to render.py
