+++ daiv/memory/management/commands/backfill_memory_entries.py
@@ -9,7 +9,8 @@ from asgiref.sync import async_to_sync
 
 from codebase.repo_config import RepositoryConfig
 from memory.models import MemoryEntry, MemoryObservation
-from memory.tasks import document_size, render_memory_document, run_consolidation_round
+from memory.render import document_size, render_memory_document
+from memory.tasks import run_consolidation_round
 
 logger = logging.getLogger("daiv.memory")
