+++ tests/unit_tests/memory/test_render.py
@@ -5,7 +5,7 @@ from django.utils import timezone
 import pytest
 from memory.constants import MEMORY_MAX_BYTES, MEMORY_MAX_LINES
 from memory.models import MemoryEntry, ObservationCategory
-from memory.tasks import CATEGORY_SECTIONS, prune_to_budget, render_memory_document
+from memory.render import CATEGORY_SECTIONS, prune_to_budget, render_memory_document
 
 NOW = timezone.now()
